requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0

# LangChain ecosystem
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7

# Azure OpenAI
openai==1.30.1

# Vector store
faiss-cpu==1.8.0

# Embeddings
sentence-transformers==3.0.0
transformers==4.41.1
torch==2.3.0

# Document loaders
pypdf==4.2.0
unstructured[docx]==0.13.4
python-docx==1.1.2

# Evaluation
ragas==0.1.9
datasets==2.19.1

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.66.4
